Rename submodules
Give them prefixes to make them more clear.
